libhsmd: Migrate handle_sign_withdrawal_tx
